Fixed a few bugs causing some tests to not being to run through

+++ src/mathlib-code.h
@@ -22,7 +22,7 @@ double multiplication(double x, double y);
 
 double division(double x, double y);
 
-double factorial(int x);
+double factorial(double x);
 
 double exponentiation(double x, int n);
